+++ spec/dummy/app/tabulatr_data/vendor_tabulatr_data.rb
@@ -1,0 +1,15 @@
+class VendorTabulatrData < Tabulatr::Data
+
+  column :name
+  column :url
+  column :active
+
+  filter :product_price_range do |relation, value|
+    relation = relation.joins(:products)
+    if value == 'low'
+      relation.group("vendors.id").having('AVG(products.price) <= 100')
+    elsif value == 'high'
+      relation.group("vendors.id").having('AVG(products.price) > 100')
+    end
+  end
+end